testFabriqueTacheCombinaisonEstNombreArmstrongValide permet de tester tout le range jusqu'à 64bits.
Il y'a du refactoring à faire. Peut etre des méthodes à déplacer.

Les résultats obtenus sont super encourageants.

+++ app/puissance10.h
@@ -8,6 +8,7 @@
 #include <stdlib.h>
 #include <math.h>
 #include <assert.h>
+#include "mathematique.h"
 
 /*! \brief Cache de puissance pour la valeur 10
 
@@ -17,7 +18,7 @@
 typedef struct cache_puissance_10
 {
 	char borne_superieure;
-	long *cache;
+	long long *cache;
 } CachePuissance10;
 
 /*! \brief Initialise le cache en calculant les valeurs de 10^0 à 10^n
@@ -26,7 +27,7 @@ void InitCachePuissance10(CachePuissance10 *cache, unsigned char borne_superieur
 
 /* \brief Retourne la puissance de 10 recherchée depuis le cache
 */
-long inline GetPuissance10(CachePuissance10 *cache, unsigned char exposant);
+long long inline GetPuissance10(CachePuissance10 *cache, unsigned char exposant);
 
 /*\brief Destructeur du cache
 */
